backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
python-multipart
python-dateutil
